Mount all dev paths into the app container
Otherwise editing these files locally, on OS X, will not be reflected in
the running app container.

Notice that mix.exs, mix.lock, assets/package* as well as
assets/yarn.lock have not been mounted into the app container. If they
change, the container needs to be rebuilt since the compilation stage
happens **before** beam & webpack get booted. Any mix.exs changes will
require the BEAM VM to be restarted, so let's go with the simplistic
approach of requiring a full re-build if any of these files change. This
might be of interest to @nickjj, since this is a different take on the
"runtime lock stomping" problem:
https://github.com/nickjj/docker-web-framework-examples/commit/48428b8ff6b856b598e4082495d72e99f2df2b67#diff-73d444f37ad1743a8ae6f58d85ca5047